requirements.txt
django
psycopg2
pandas
xlrd
requests
python-gitlab
beautifulsoup4
html5lib